make use of click.edit()
